add TestUtils.cpp
