+++ tests/unit/TestUtils.hpp
@@ -23,66 +23,21 @@
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
 
-using testing::Matcher, testing::AllOf, testing::Field, testing::Property, testing::Pointwise,
-      testing::FloatEq, testing::NanSensitiveFloatEq, testing::NanSensitiveFloatNear;
+using ::testing::Matcher;
 
-Matcher<yagit::DataSize> matchDataSize(const yagit::DataSize& expectedDataSize){
-    return AllOf(Field("frames", &yagit::DataSize::frames, expectedDataSize.frames),
-                 Field("rows", &yagit::DataSize::rows, expectedDataSize.rows),
-                 Field("columns", &yagit::DataSize::columns, expectedDataSize.columns));
-}
-
-Matcher<yagit::DataOffset> matchDataOffset(const yagit::DataOffset& expectedDataOffset){
-    return AllOf(Field("frames", &yagit::DataOffset::frames, FloatEq(expectedDataOffset.frames)),
-                 Field("rows", &yagit::DataOffset::rows, FloatEq(expectedDataOffset.rows)),
-                 Field("columns", &yagit::DataOffset::columns, FloatEq(expectedDataOffset.columns)));
-}
-
-Matcher<yagit::DataSpacing> matchDataSpacing(const yagit::DataSpacing& expectedDataSpacing){
-    return AllOf(Field("frames", &yagit::DataSpacing::frames, FloatEq(expectedDataSpacing.frames)),
-                 Field("rows", &yagit::DataSpacing::rows, FloatEq(expectedDataSpacing.rows)),
-                 Field("columns", &yagit::DataSpacing::columns, FloatEq(expectedDataSpacing.columns)));
-}
-
-Matcher<yagit::ImageData> matchImageData(const yagit::ImageData& expectedImageData, float maxAbsError = -1){
-    return AllOf(Property("data", &yagit::ImageData::getData,
-                          Pointwise((maxAbsError <= 0 ? NanSensitiveFloatEq() : NanSensitiveFloatNear(maxAbsError)),
-                                    expectedImageData.getData())),
-                 Property("size", &yagit::ImageData::getSize, matchDataSize(expectedImageData.getSize())),
-                 Property("offset", &yagit::ImageData::getOffset, matchDataOffset(expectedImageData.getOffset())),
-                 Property("spacing", &yagit::ImageData::getSpacing, matchDataSpacing(expectedImageData.getSpacing())));
-}
-
+Matcher<yagit::DataSize> matchDataSize(const yagit::DataSize& expectedDataSize);
+Matcher<yagit::DataOffset> matchDataOffset(const yagit::DataOffset& expectedDataOffset);
+Matcher<yagit::DataSpacing> matchDataSpacing(const yagit::DataSpacing& expectedDataSpacing);
+Matcher<yagit::ImageData> matchImageData(const yagit::ImageData& expectedImageData, float maxAbsError = -1);
 Matcher<yagit::ImageData> matchImageData(const std::vector<float>& expectedData,
                                          const yagit::DataSize& expectedDataSize,
                                          const yagit::DataOffset& expectedDataOffset,
                                          const yagit::DataSpacing& expectedDataSpacing,
-                                         float maxAbsError = -1){
-    return AllOf(Property("data", &yagit::ImageData::getData,
-                          Pointwise((maxAbsError <= 0 ? NanSensitiveFloatEq() : NanSensitiveFloatNear(maxAbsError)),
-                                    expectedData)),
-                 Property("size", &yagit::ImageData::getSize, matchDataSize(expectedDataSize)),
-                 Property("offset", &yagit::ImageData::getOffset, matchDataOffset(expectedDataOffset)),
-                 Property("spacing", &yagit::ImageData::getSpacing, matchDataSpacing(expectedDataSpacing)));
-}
-
+                                         float maxAbsError = -1);
 
 yagit::ImageData generateImageData(float value, const yagit::DataSize& size,
-                                   const yagit::DataOffset& offset, const yagit::DataSpacing& spacing){
-    size_t count = size.frames * size.rows * size.columns;
-    return yagit::ImageData(std::vector<float>(count, value), size, offset, spacing);
-}
-
+                                   const yagit::DataOffset& offset, const yagit::DataSpacing& spacing);
 
 namespace yagit{
-std::ostream& operator<<(std::ostream& os, const yagit::GammaParameters& gammaParams){
-    std::string norm = (gammaParams.normalization == yagit::GammaNormalization::Global ? "global" : (
-        gammaParams.normalization == yagit::GammaNormalization::Local ? "local" : "?"
-    ));
-    return os << "{dd=" << gammaParams.ddThreshold << ", dta=" << gammaParams.dtaThreshold
-              << ", norm=" << norm << ", global norm dose=" << gammaParams.globalNormDose
-              << ", dco=" << gammaParams.doseCutoff
-              << ", max search dist=" << gammaParams.maxSearchDistance
-              << ", step size=" << gammaParams.stepSize << "}";
-}
+std::ostream& operator<<(std::ostream& os, const yagit::GammaParameters& gammaParams);
 }
